requirements.txt
# Core data processing and visualization
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Machine learning
scikit-learn>=1.0.0
scipy>=1.7.0

# Web application framework
streamlit>=1.10.0

# API requests
requests>=2.25.0

# Optional visualization enhancements
plotly>=5.3.0
altair>=4.1.0

# Data export
openpyxl>=3.0.0  # For Excel export support
xlsxwriter>=3.0.0

# Environment management
python-dotenv>=0.19.0  # For managing API keys

# For Jupyter Notebook development (optional)
jupyterlab>=3.0.0